+++ PID_v1.h
@@ -15,11 +15,11 @@ public:
 #define P_ON_E 1
 
   //commonly used functions **************************************************************************
-    PID(long*, double*, long*,        // * constructor.  links the PID to the Input, Output, and
+    PID(double*, double*, double*,        // * constructor.  links the PID to the Input, Output, and
         double, double, double, int, int);//   Setpoint.  Initial tuning parameters are also set here.
                                           //   (overload for specifying proportional mode)
 
-    PID(long*, double*, long*,        // * constructor.  links the PID to the Input, Output, and
+    PID(double*, double*, double*,        // * constructor.  links the PID to the Input, Output, and
         double, double, double, int);     //   Setpoint.  Initial tuning parameters are also set here
 
     void SetMode(int Mode);               // * sets PID to either Manual (0) or Auto (non-0)
@@ -73,9 +73,9 @@ private:
     int controllerDirection;
     int pOn;
 
-    long *myInput;              // * Pointers to the Input, Output, and Setpoint variables
+    double *myInput;              // * Pointers to the Input, Output, and Setpoint variables
     double *myOutput;             //   This creates a hard link between the variables and the
-    long *mySetpoint;           //   PID, freeing the user from having to constantly tell us
+    double *mySetpoint;           //   PID, freeing the user from having to constantly tell us
                                   //   what these values are.  with pointers we'll just know.
 
     unsigned long lastTime;
